Rework docker image
Build the package from source, this will spear us the wait for the package
to be available from the official repositories to make the package
available and mixing alpine repositories.

To achive this we use the multi-stage build feature of docker to do the
actual package building in a "builder" stage and copy the resulting
package to the "main" stage.